Card class moved to its own file
